+++ CHANGELOG.rst
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+0.2.0 (xxxx-xx-xx)
+------------------
+
+Move requirements to motor>=2.0.
+
+Remove ``count`` method from ``MotorQuerySetCursor``.
+
+Add session support to ``MotorQuerySet`` and ``Document``.
+
+Add ``create_collection`` method to ``Document``.
 
 0.1.0 (2017-05-19)
 ------------------